requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
Pillow==10.4.0
psutil==6.0.0
pyparsing==3.1.4
pystray==0.19.5
python-dateutil==2.9.0.post0
six==1.16.0
